+++ src/main/java/net/meander/subtlyd/mixin/common/entity/ai/goal/NearestAttackableTargetGoalMixin.java
@@ -1,0 +1,41 @@
+package net.meander.subtlyd.mixin.common.entity.ai.goal;
+
+import net.meander.subtlyd.util.Util;
+import net.meander.subtlyd.util.data.tags.EntityTypeTagsSD;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
+import net.minecraft.world.entity.animal.Animal;
+import org.jspecify.annotations.Nullable;
+import org.spongepowered.asm.mixin.Final;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(NearestAttackableTargetGoal.class)
+public class NearestAttackableTargetGoalMixin {
+    @Shadow @Final protected Class<?> targetType;
+    @Shadow @Nullable protected LivingEntity target;
+
+    /**
+     * Determines when hunting cooldowns should be used.
+     */
+    @Inject(method = "canUse", at = @At("HEAD"), cancellable = true)
+    private void checkHuntCooldowns(CallbackInfoReturnable<Boolean> cir) {
+        Mob mob = ((TargetGoalAccessor) this).subtlyDungeons$getMob();
+
+        if (mob.is(EntityTypeTagsSD.CAN_BE_FULL) && Animal.class.isAssignableFrom(targetType)) {
+            Long nextAllowedHuntTime = Util.Logic.HUNT_COOLDOWNS.get(mob);
+
+            if (nextAllowedHuntTime != null && mob.level().getGameTime() < nextAllowedHuntTime) {
+                cir.setReturnValue(false);
+            }
+        }
+
+        if (mob.is(EntityTypeTagsSD.NOCTURNAL_HUNTER) && mob.level().isBrightOutside()) {
+            cir.setReturnValue(false);
+        }
+    }
+}